requirements.txt
babel-2.12.1 
django-money-3.0.0 
py-moneyed-2.0 
typing-extensions-4.5.0